refactor: Minor refactoring
Resolves a few minor issues in the code (like missing parenthesis for
code blocks, trailing commas, or bad names), restructures a few test
cases from separate "GitHubIssue" files to where they semantically
belong to, and reworks some tests to e.g. actually test something or
remove `println` statements.

This does not solve all similar issues, and does not attempt to.
