+++ kgraphql/src/main/kotlin/com/apurebase/kgraphql/schema/structure/Type.kt
@@ -66,7 +66,7 @@ interface Type : __Type {
     val kClass: KClass<*>?
 
     abstract class ComplexType(val allFields: List<Field>) : Type {
-        val fieldsByName = allFields.associateBy { it.name }
+        private val fieldsByName = allFields.associateBy { it.name }
 
         override val fields: List<__Field>? = allFields.filterNot { it.name.startsWith("__") }
