+++ kgraphql/src/test/kotlin/com/apurebase/kgraphql/specification/language/FragmentsSpecificationTest.kt
@@ -2,6 +2,7 @@ package com.apurebase.kgraphql.specification.language
 
 import com.apurebase.kgraphql.Actor
 import com.apurebase.kgraphql.GraphQLError
+import com.apurebase.kgraphql.KGraphQL
 import com.apurebase.kgraphql.Specification
 import com.apurebase.kgraphql.assertNoErrors
 import com.apurebase.kgraphql.defaultSchema
@@ -11,6 +12,7 @@ import com.apurebase.kgraphql.extract
 import com.apurebase.kgraphql.integration.BaseSchemaTest
 import com.apurebase.kgraphql.request.Introspection
 import org.amshove.kluent.invoking
+import org.amshove.kluent.shouldBeEqualTo
 import org.amshove.kluent.shouldThrow
 import org.amshove.kluent.withMessage
 import org.hamcrest.CoreMatchers.equalTo
@@ -188,5 +190,148 @@ class FragmentsSpecificationTest {
         """
             )
         } shouldThrow GraphQLError::class withMessage "There can be only one fragment named film_title."
+    }
+
+    sealed class TopUnion(val field: String) {
+        class Union1(val names: List<String>) : TopUnion("union1")
+        class Union2(val numbers: List<Int>) : TopUnion("union2")
+    }
+
+    // https://github.com/aPureBase/KGraphQL/issues/141
+    @Test
+    fun `fragments on union types should work`() {
+        KGraphQL.schema {
+            unionType<TopUnion>()
+
+            query("unions") {
+                resolver { isOne: Boolean ->
+                    if (isOne) {
+                        TopUnion.Union1(listOf("name1", "name2"))
+                    } else {
+                        TopUnion.Union2(listOf(1, 2))
+                    }
+                }
+            }
+        }.executeBlocking(
+            """
+            {
+                unions(isOne: true) {
+                    ...abc
+                }
+            }
+            fragment abc on TopUnion {
+                ... on Union1 { names }
+                ... on Union2 { numbers }
+            }
+            """.trimIndent()
+        )
+            .deserialize()
+            .extract<List<String>>("data/unions/names") shouldBeEqualTo listOf("name1", "name2")
+    }
+
+    data class Outer(val inner1: Inner, val inner2: Inner)
+    data class Inner(val name: String)
+
+    private val testedSchema = defaultSchema {
+        query("outer") {
+            resolver { ->
+                Outer(Inner("test1"), Inner("test2"))
+            }
+        }
+        type<Inner> {
+            property("testProperty1") {
+                resolver { "${it.name}.testProperty1" }
+            }
+            property("testProperty2") {
+                resolver { "${it.name}.testProperty2" }
+            }
+        }
+    }
+
+    // https://github.com/aPureBase/KGraphQL/issues/197
+    @Test
+    fun `executor should merge fragment declaration and field declaration`() {
+        val response = testedSchema.executeBlocking(
+            """
+            { 
+                outer { 
+                    ...TestFragment
+                    inner1 { testProperty1 } 
+                    inner2 { testProperty2 } 
+                } 
+            } 
+            fragment TestFragment on Outer {
+                inner1 { name, testProperty2 }
+                inner2 { name, testProperty1 }
+            }
+        """.trimIndent()
+        )
+
+        val deserialized = deserialize(response)
+        assertThat(
+            deserialized, equalTo(
+                mapOf(
+                    "data" to mapOf(
+                        "outer" to mapOf(
+                            "inner1" to mapOf(
+                                "name" to "test1",
+                                "testProperty2" to "test1.testProperty2",
+                                "testProperty1" to "test1.testProperty1",
+                            ),
+                            "inner2" to mapOf(
+                                "name" to "test2",
+                                "testProperty1" to "test2.testProperty1",
+                                "testProperty2" to "test2.testProperty2",
+                            ),
+                        )
+                    )
+                )
+            )
+        )
+    }
+
+    // https://github.com/aPureBase/KGraphQL/issues/197
+    @Test
+    fun `executor should merge several fragment declarations and field declaration`() {
+        val response = testedSchema.executeBlocking(//language=graphql
+            """
+            { 
+                outer { 
+                    ...TestFragment1
+                    ...TestFragment2
+                    inner2 {
+                        testProperty2
+                    }
+                } 
+            } 
+            fragment TestFragment1 on Outer {
+                inner1 { name, testProperty2 }
+            }
+            fragment TestFragment2 on Outer {
+                inner2 { name, testProperty1 }
+            }
+        """.trimIndent()
+        )
+
+        val deserialized = deserialize(response)
+        assertThat(
+            deserialized, equalTo(
+                mapOf(
+                    "data" to mapOf(
+                        "outer" to mapOf(
+                            "inner1" to mapOf(
+                                "name" to "test1",
+                                "testProperty2" to "test1.testProperty2",
+                            ),
+                            "inner2" to mapOf(
+                                "name" to "test2",
+                                "testProperty1" to "test2.testProperty1",
+                                "testProperty2" to "test2.testProperty2",
+                            ),
+                        )
+                    )
+                )
+            )
+        )
     }
 }
